refactor: 상수들을 리팩터링
