+++ frontend/src/components/Template/CustomerTemplate/index.tsx
@@ -2,8 +2,8 @@ import { PropsWithChildren } from 'react';
 import { BaseCustomerTemplate, ContentContainer } from './style';
 import BottomTabBar from './BottomTabBar';
 import { useLocation } from 'react-router-dom';
-import { ROUTER_PATH } from '../../../constants';
 import { RouterPath } from '../../../types/utils';
+import ROUTER_PATH from '../../../constants/routerPath';
 
 const CustomerTemplate = ({ children }: PropsWithChildren) => {
   const location = useLocation();
