+++ frontend/src/pages/Admin/CouponDesign/ChoiceTemplate/index.tsx
@@ -1,13 +1,28 @@
 import { Dispatch, SetStateAction, useState } from 'react';
 import { ChoiceTemplateContainer } from './style';
 import TabBar from '../../../../components/TabBar';
-import { TEMPLATE_MENU, TEMPLATE_OPTIONS } from '../../../../constants';
+import { TEMPLATE_MENU } from '../../../../constants';
 import {
   SampleBackCouponImage,
   SampleImage,
   StampCoordinate,
 } from '../../../../types/domain/coupon';
 import SampleImageList from './SampleImageList';
+
+export const TEMPLATE_OPTIONS = [
+  {
+    key: 'coupon-front',
+    value: TEMPLATE_MENU.FRONT_IMAGE,
+  },
+  {
+    key: 'coupon-back',
+    value: TEMPLATE_MENU.BACK_IMAGE,
+  },
+  {
+    key: 'stamp',
+    value: TEMPLATE_MENU.STAMP,
+  },
+];
 
 interface ChoiceTemplateProps {
   frontImageUrl: string;
