+++ frontend/src/pages/Admin/EarnStamp/hooks/usePostEarnStamp.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { postEarnStamp } from '../../../../api/post';
 import { useNavigate } from 'react-router-dom';
-import { ROUTER_PATH } from '../../../../constants';
+import ROUTER_PATH from '../../../../constants/routerPath';
 
 const usePostEarnStamp = () => {
   const navigate = useNavigate();
