+++ frontend/src/pages/Admin/ManageCafe/components/CafePhoneNumber/index.tsx
@@ -1,5 +1,5 @@
 import { Input } from '../../../../../components/Input';
-import { PHONE_NUMBER_REGEX } from '../../../../../constants';
+import { REGEX } from '../../../../../constants/regex';
 import { parsePhoneNumber } from '../../../../../utils';
 import { StepTitle, Wrapper } from '../../style';
 
@@ -16,7 +16,7 @@ const CafePhoneNumber = ({ phoneNumber, inputPhoneNumber }: CafePhoneNumberProps
         id="phone-number-input"
         placeholder="전화번호를 입력해주세요"
         maxLength={phoneNumber.startsWith('02') ? 12 : 13}
-        pattern={PHONE_NUMBER_REGEX.source}
+        pattern={REGEX.phoneNumber.source}
         onChange={inputPhoneNumber}
         value={parsePhoneNumber(phoneNumber)}
       />
