+++ frontend/src/pages/Auth/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import { getOAuthToken } from '../../api/get';
-import { ROUTER_PATH } from '../../constants';
+import ROUTER_PATH from '../../constants/routerPath';
 
 const Auth = () => {
   const navigate = useNavigate();
