+++ frontend/src/pages/Customer/CouponList/components/Header/index.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
-import { ROUTER_PATH } from '../../../../../constants';
 import { HeaderContainer } from '../../style';
 import { StampcrushLogo } from '../../../../../assets';
+import ROUTER_PATH from '../../../../../constants/routerPath';
 
 const Header = () => {
   return (
